Organize the MapModel creator's code, moving demand data management to a module
